requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dm-tree==0.1.8
Flask==3.0.2
flatbuffers==24.3.7
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h5py   
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes    
namex==0.0.7
numpy==1.26.4   
opt-einsum  
packaging==24.0
pillow==10.2.0
protobuf==3.20.3
Pygments==2.17.2
requests==2.31.0
rich==13.7.1
setuptools==69.2.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server
tensorflow==2.16.1
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0